requirements.txt
Django==4.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-writable-nested==0.6.3
Pillow==10.4.0
pyotp==2.9.0
celery==5.4.0
redis==5.0.8
django-celery-beat==2.5.0
django-celery-results==2.4.0
django-channels==0.7.0
requests==2.32.3
paho-mqtt==2.1.0
joblib==1.4.2
numpy==2.0.1
scikit-learn==1.5.1